+++ Classes/Code/TagManagerCodeBuilder.php
@@ -25,7 +25,6 @@ class TagManagerCodeBuilder
 {
     private ServerRequestInterface $request;
     private Configuration $configuration;
-    private EventDispatcherInterface $eventDispatcher;
     /**
      * @var JavaScriptCode[]
      */
@@ -35,9 +34,9 @@ class TagManagerCodeBuilder
      */
     private array $dataLayerVariables = [];
 
-    public function __construct(EventDispatcherInterface $eventDispatcher)
-    {
-        $this->eventDispatcher = $eventDispatcher;
+    public function __construct(
+        private readonly EventDispatcherInterface $eventDispatcher
+    ) {
     }
 
     public function setRequest(ServerRequestInterface $request): self