+++ Tests/Unit/Code/MatomoMethodCallTest.php
@@ -14,7 +14,6 @@ namespace Brotkrueml\MatomoIntegration\Tests\Unit\Code;
 use Brotkrueml\MatomoIntegration\Code\JavaScriptCode;
 use Brotkrueml\MatomoIntegration\Code\MatomoMethodCall;
 use Brotkrueml\MatomoIntegration\Exceptions\InvalidMatomoMethodName;
-use Brotkrueml\MatomoIntegration\Exceptions\InvalidMatomoMethodParameter;
 use PHPUnit\Framework\TestCase;
 
 final class MatomoMethodCallTest extends TestCase
@@ -115,17 +114,5 @@ final class MatomoMethodCallTest extends TestCase
         $this->expectExceptionMessage('The given Matomo method name "some method name" is not valid, only characters between a and z are allowed!');
 
         new MatomoMethodCall('some method name');
-    }
-
-    /**
-     * @test
-     */
-    public function exceptionIsThrownWhenInvalidParameterTypeIsGiven(): void
-    {
-        $this->expectException(InvalidMatomoMethodParameter::class);
-        $this->expectExceptionCode(1629212630);
-        $this->expectExceptionMessage('A Matomo method argument with the invalid type "null" was given, allowed: array, bool, int, string, Brotkrueml\\MatomoIntegration\\Code\\JavaScriptCode');
-
-        (new MatomoMethodCall('someMethodName', null))->__toString();
     }
 }