+++ ext_emconf.php
@@ -10,7 +10,7 @@ $EM_CONF[$_EXTKEY] = [
     'constraints' => [
         'depends' => [
             'typo3' => '11.5.0-12.4.99',
-            'php' => '7.4.0-0.0.0',
+            'php' => '8.1.0-0.0.0',
         ],
         'conflicts' => [],
         'suggests' => [],