+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 - Compatibility with TYPO3 v10 (#17)
+- Compatibility with PHP 7.4 and 8.0 (#18)
 
 ## [1.8.0] - 2023-11-19